Change tab bar for one view
