+++ TheMovieDb/Models/MovieManager.swift
@@ -12,7 +12,6 @@ class MovieManager{
     let apiKey = "?api_key=444cd656b00475d785aa41a9c43b2e44"
     var apiService = APIService()
     var popularMovies = [Movie]()
-    //var delegate: MovieManagerDelegate?
     
     func fetchMoviesData(type: String, completion: @escaping () -> ()) {
     let urlString = "\(movieBaseURL)\(type)\(apiKey)"
@@ -25,11 +24,10 @@ class MovieManager{
                 completion()
             case .failure(let error):
                 // Something is wrong with the JSON file or the model
-                print("Error processing json data: \(error)")
+                print("Error processing JSON data: \(error)")
             }
         }
         
-        //performRequest(with: urlString)
     }
     
     func numberOfRowsInSection(section: Int) -> Int {
